+++ tests/test_training.py
@@ -1,17 +1,17 @@
 import subprocess
 
 from pathlib import Path
-from app.src.training import Training
+from ml_pipeline.src.training import Training
 
 import pytest
 import sys
 
-from app.utils.enumerations import Directory, MachineLearningModels
+from ml_pipeline.utils.enumerations import Directory, MachineLearningModels
 
 
 def test_save_model(training: Training, tmp_path: Path):
     save_model_dir = tmp_path / 'random_forest_train.joblib'
-    arguments = [sys.executable, '-m', 'scripts.train_model',
+    arguments = [sys.executable, '-m', 'ml_pipeline.train_model',
                  '--raw_data_dir', str(Directory.IRIS_DATA_DIR.value),
                  '--model', MachineLearningModels.random_forest_default.name,
                  '--model_save_dir', str(save_model_dir)]
